requirements.txt
python-telegram-bot==20.7
asyncio
uuid
json
os
time
datetime